pip-requirements.txt
py2exe-py2==0.6.9
requests==2.12.4